Rename DeadlineInput
