+++ src/cow-react/modules/limitOrders/containers/LimitOrdersWidget/index.tsx
@@ -13,7 +13,7 @@ import { useAtomValue, useUpdateAtom } from 'jotai/utils'
 import { SettingsWidget } from '../SettingsWidget'
 import { limitOrdersSettingsAtom } from '../../state/limitOrdersSettingsAtom'
 import { RateInput } from '../RateInput'
-import { ExpiryDate } from '../ExpiryDate'
+import { ExpiryDate } from '../DeadlineInput'
 import { useUpdateCurrencyAmount } from '../../hooks/useUpdateCurrencyAmount'
 import { LimitOrdersConfirmModal } from '../LimitOrdersConfirmModal'
 import { limitOrdersQuoteAtom } from '../../state/limitOrdersQuoteAtom'
